Add (dumb) check for apt-cacher-ng
apt-cacher-ng does not support HTTPS so if the string '3142' (the
default apt-cacher-ng port) is found in /etc/apt/apt.conf*, we'll add
the Tor repository as HTTP.

Perhaps this check could/should be extended to check whether the
torproject repo was added to apt-cacher-ng's `PassThroughPattern`
setting. Or maybe we could take a stab at updating that ourselves (I
think I'd prefer *advising* instead).